Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float n;
-    scanf("%f",&n);
+    double n;
+    scanf("%lf",&n);
     printf("You entered: %.lf",n);
     return 0;
 }
